+++ experiment/src/org/labkey/experiment/ExperimentModule.java
@@ -54,6 +54,7 @@ import org.labkey.api.exp.api.ExpSampleType;
 import org.labkey.api.exp.api.ExperimentJSONConverter;
 import org.labkey.api.exp.api.ExperimentService;
 import org.labkey.api.exp.api.FilterProtocolInputCriteria;
+import org.labkey.api.exp.api.ExpLineageService;
 import org.labkey.api.exp.api.SampleTypeDomainKind;
 import org.labkey.api.exp.api.SampleTypeService;
 import org.labkey.api.exp.api.StorageProvisioner;
@@ -110,7 +111,7 @@ import org.labkey.experiment.api.ExpSampleTypeImpl;
 import org.labkey.experiment.api.ExperimentServiceImpl;
 import org.labkey.experiment.api.ExperimentStressTest;
 import org.labkey.experiment.api.GraphAlgorithms;
-import org.labkey.experiment.api.LineagePerfTest;
+import org.labkey.experiment.lineage.LineagePerfTest;
 import org.labkey.experiment.api.LineageTest;
 import org.labkey.experiment.api.LogDataType;
 import org.labkey.experiment.api.Protocol;
@@ -133,6 +134,7 @@ import org.labkey.experiment.api.property.TextChoiceValidator;
 import org.labkey.experiment.controllers.exp.ExperimentController;
 import org.labkey.experiment.controllers.property.PropertyController;
 import org.labkey.experiment.defaults.DefaultValueServiceImpl;
+import org.labkey.experiment.lineage.ExpLineageServiceImpl;
 import org.labkey.experiment.pipeline.ExperimentPipelineProvider;
 import org.labkey.experiment.samples.DataClassFolderImporter;
 import org.labkey.experiment.samples.DataClassFolderWriter;
@@ -198,6 +200,7 @@ public class ExperimentModule extends SpringModule
         SampleTypeService.setInstance(new SampleTypeServiceImpl());
         DefaultValueService.setInstance(new DefaultValueServiceImpl());
         StorageProvisioner.setInstance(StorageProvisionerImpl.get());
+        ExpLineageService.setInstance(new ExpLineageServiceImpl());
 
         PropertyServiceImpl propertyServiceImpl = new PropertyServiceImpl();
         PropertyService.setInstance(propertyServiceImpl);
