+++ experiment/src/org/labkey/experiment/controllers/exp/ExperimentController.java
@@ -159,7 +159,6 @@ import org.labkey.api.util.ErrorRenderer;
 import org.labkey.api.util.ExceptionUtil;
 import org.labkey.api.util.FileStream;
 import org.labkey.api.util.FileUtil;
-import org.labkey.api.util.GUID;
 import org.labkey.api.util.HtmlString;
 import org.labkey.api.util.HtmlStringBuilder;
 import org.labkey.api.util.ImageUtil;
@@ -217,6 +216,7 @@ import org.labkey.experiment.api.ProtocolActionStepDetail;
 import org.labkey.experiment.api.SampleTypeServiceImpl;
 import org.labkey.experiment.api.SampleTypeUpdateServiceDI;
 import org.labkey.experiment.controllers.property.PropertyController;
+import org.labkey.experiment.lineage.ExpLineageServiceImpl;
 import org.labkey.experiment.pipeline.ExperimentPipelineJob;
 import org.labkey.experiment.types.TypesController;
 import org.labkey.experiment.xar.XarExportSelection;
@@ -7004,131 +7004,14 @@ public class ExperimentController extends SpringActionController
         }
     }
 
-
-    @RequiresPermission(ReadPermission.class)
-    public static class LineageOldAction extends BaseResolveLsidApiAction<ExpLineageOptions>
-    {
-        @Override
-        public Object execute(ExpLineageOptions options, BindException errors)
-        {
-            ExpLineage lineage = ExperimentServiceImpl.get().getLineage(getContainer(), getUser(), _seeds, options);
-            var settings = new ExperimentJSONConverter.Settings(options.isIncludeProperties(), options.isIncludeInputsAndOutputs(), options.isIncludeRunSteps());
-            return new ApiSimpleResponse(lineage.toJSON(getUser(), options.isSingleSeedRequested(), settings));
-        }
-    }
-
     @RequiresPermission(ReadPermission.class)
     public static class LineageAction extends BaseResolveLsidApiAction<ExpLineageOptions>
     {
-        private static final int BATCH_SIZE = 5_000;
-
-        private record Context(
-            User user,
-            ApiJsonWriter writer,
-            ExperimentJSONConverter.Settings settings,
-            Map<String, ExpLineage.Edges> nodesAndEdges,
-            Map<GUID, Boolean> hasPermission
-        )
-        {
-            boolean hasPermission(Container container, User user)
-            {
-                return this.hasPermission.computeIfAbsent(container.getEntityId(), (id) -> container.hasPermission(user, ReadPermission.class));
-            }
-
-            @NotNull ExpLineage.Edges popEdges(String lsid)
-            {
-                var edges = this.nodesAndEdges.get(lsid);
-                this.nodesAndEdges.remove(lsid);
-                return edges == null ? ExpLineage.Edges.emptyEdges : edges;
-            }
-        }
-
         @Override
         public Object execute(ExpLineageOptions options, BindException errors) throws Exception
         {
-            var lineage = ExperimentServiceImpl.get().getLineageResult(getContainer(), getUser(), _seeds, options);
-
-            var context = new Context(
-                getUser(),
-                new ApiJsonWriter(getViewContext().getResponse()),
-                new ExperimentJSONConverter.Settings(options.isIncludeProperties(), options.isIncludeInputsAndOutputs(), options.isIncludeRunSteps()),
-                ExpLineage.processEdges(lineage.edges()),
-                new HashMap<>()
-            );
-
-            context.writer.startResponse();
-
-            writeNodes(lineage, context);
-            writeSeed(lineage, context, options);
-
-            context.writer.endResponse();
-
+            ExpLineageServiceImpl.get().streamLineage(getContainer(), getUser(), getViewContext().getResponse(), _seeds, options);
             return null;
-        }
-
-        private static void writeNodes(ExperimentServiceImpl.LineageResult lineage, Context context) throws IOException
-        {
-            context.writer.startObject("nodes");
-
-            if (!context.nodesAndEdges.isEmpty())
-            {
-                if (!lineage.dataIds().isEmpty())
-                    ExperimentServiceImpl.get().forEachBatchExpData(lineage.dataIds(), BATCH_SIZE, writeBatch(context));
-                if (!lineage.materialIds().isEmpty())
-                    ExperimentServiceImpl.get().forEachBatchExpMaterial(lineage.materialIds(), BATCH_SIZE, writeBatch(context));
-                if (!lineage.runIds().isEmpty())
-                    ExperimentServiceImpl.get().forEachBatchExpRun(lineage.runIds(), BATCH_SIZE, writeBatch(context));
-                if (!lineage.objectLsids().isEmpty())
-                {
-                    var lsidManager = LsidManager.get();
-                    writeBatchList(lineage.objectLsids().stream().map(lsidManager::getObject).toList(), context);
-                }
-            }
-
-            writeBatchList(lineage.seeds().stream().toList(), context);
-
-            for (var entry : context.nodesAndEdges.entrySet())
-                context.writer.writeProperty(entry.getKey(), nodeToJson(null, entry.getValue(), context));
-
-            context.writer.endObject();
-        }
-
-        private static void writeSeed(ExperimentServiceImpl.LineageResult lineage, Context context, ExpLineageOptions options) throws IOException
-        {
-            // If the request was made with a single 'seed' property, use single 'seed' property in the response
-            // otherwise, include an array of 'seed' regardless of the number of seed items.
-            if (options.isSingleSeedRequested())
-                context.writer.writeProperty("seed", lineage.seeds().stream().findFirst().orElseThrow().getLSID());
-            else
-                context.writer.writeProperty("seeds", lineage.seeds().stream().map(Identifiable::getLSID).toList());
-        }
-
-        private static Selector.ForEachBatchBlock writeBatch(Context context)
-        {
-            return (batch) -> {
-                try
-                {
-                    writeBatchList(batch, context);
-                }
-                catch (IOException e)
-                {
-                    throw new RuntimeException(e);
-                }
-            };
-        }
-
-        private static void writeBatchList(List<? extends Identifiable> batch, Context context) throws IOException
-        {
-            for (var node : batch)
-            {
-                if (context.hasPermission(node.getContainer(), context.user))
-                    context.writer.writeProperty(node.getLSID(), nodeToJson(node, context.popEdges(node.getLSID()), context));
-            }
-        }
-
-        private static JSONObject nodeToJson(@Nullable Identifiable node, ExpLineage.Edges edges, Context context)
-        {
-            return ExpLineage.nodeToJSON(node, context.user, edges, context.settings);
         }
     }
